feat(contract): consolidate BuildStep/BuildEventKind into @gamedevpl/contract
north-star-architecture.md Phase 1, fourth slice. BuildStep (eight
build-progress steps) and BuildEventKind (five event kinds) were each
hand-duplicated as identical string unions in apps/api/src/submission-status.ts
and apps/web/src/submissionApi.ts, same shape as SubmissionState before it.

Moved both into packages/contract/src/build-event.ts as const arrays with
derived types. submission-status.ts now re-exports the contract's values
(its own two importers, agent-channel.ts and mcp-server.ts, still pull
BUILD_STEPS/BUILD_EVENT_KINDS from it unchanged); submissionApi.ts imports
and re-exports the types directly.
